various reworks, simplify tests, add benchmarks (#38)

+++ loony/node.nim
@@ -76,7 +76,7 @@ template prn*(idx: uint16): uint16 =
   when loonyRotate:
     # multiply by cacheLineSize, mod by loonySlotCount
     # then add idx*cacheLineSize/loonySlotCount
-    (idx shl lShiftBits) and (loonySlotCount - 1) or (idx shr rShiftBits)
+    ((idx shl lShiftBits) and uint16(loonySlotCount - 1)) or (idx shr rShiftBits)
   else:
     idx
 
@@ -102,6 +102,7 @@ proc prepareElement*[T](el: var T): uint =
       if owners != 0:
         raise AssertionDefect.newException:
           "pushed ref shared by " & $(1 + owners) & " owners"
+    # Don't increment RC here - wasMoved prevents the source from decrementing
   result = cast[uint](el) or WRITER
   wasMoved el
